renaming of package in the maps
