+++ data-structures/maps/mutablemaps/README.md
@@ -58,4 +58,4 @@ The SMT queries generated by Stainless during the verification of the `MutableLo
 
 To run the main class, run in `sbt`
 
-```$ runMain ch.epfl.chassot.Main```
+```$ runMain ch.epfl.map.Main```
