+++ data-structures/maps/mutablemaps/src/main/scala/ch/epfl/map/EfficientFill.scala
@@ -1,7 +1,7 @@
 /** This file contains implementations of a "fill" function for Array, used as benchmark.
   */
 
-package ch.epfl.chassot
+package ch.epfl.map
 
 import MutableLongMap.*
