+++ data-structures/maps/mutablemaps/src/main/scala/ch/epfl/map/MutableLongMap.scala
@@ -1,6 +1,6 @@
 /** Author: Samuel Chassot
   */
-package ch.epfl.chassot
+package ch.epfl.map
 
 import stainless.annotation._
 import stainless.collection._
