+++ data-structures/maps/mutablemaps/src/main/scala/ch/epfl/map/OptimisedChecks.scala
@@ -1,7 +1,7 @@
 /** Author: Samuel Chassot
   */
 
-package ch.epfl.chassot
+package ch.epfl.map
 
 object OptimisedChecks {
   extension [T](inline value: T) inline def ensuring(condition: T => Boolean): T = value
