+++ data-structures/maps/mutablemaps/verify.sh
@@ -1,7 +1,7 @@
 stainless-dotty\
- ./src/main/scala/ch/epfl/chassot/ListMap.scala\
- ./src/main/scala/ch/epfl/chassot/ListLongMap.scala\
- ./src/main/scala/ch/epfl/chassot/MutableLongMap.scala\
- ./src/main/scala/ch/epfl/chassot/MutableHashMap.scala\
- ./src/main/scala/ch/epfl/chassot/iMutableMaps.scala\
+ ./src/main/scala/ch/epfl/map/ListMap.scala\
+ ./src/main/scala/ch/epfl/map/ListLongMap.scala\
+ ./src/main/scala/ch/epfl/map/MutableLongMap.scala\
+ ./src/main/scala/ch/epfl/map/MutableHashMap.scala\
+ ./src/main/scala/ch/epfl/map/iMutableMaps.scala\
   --config-file=stainless.conf -Dparallel=6 $1
